attendance_system-main/requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
Pillow==10.0.1
qrcode[pil]==7.4.2
pytesseract==0.3.10
pandas==2.1.1
numpy==1.24.3
werkzeug==2.3.7
html5-qrcode==2.3.8